Update config.yaml with new variables for dotfiles setup
Username is needed for the directory in which the dotfiles will be
installed. The various install_* options install the "groups" of
packages for the selected option. Theoretically all can be set to false
and the packages can be manually overridden by defining a yaml in
config.yaml as config_wayland: ....